requirements.txt
amqp==5.0.9
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-debug-toolbar==3.2.4
django-widget-tweaks==1.4.12
idna==3.3
Khayyam==3.0.17
kombu==5.2.3
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.4.1
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
djangorestframework
django-filter
dj_rest_auth
django-allauth
djangorestframework-simplejwt
django-decouple==2.0
django-ckeditor
elasticsearch==7.16.2
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl==7.2.2
branca==0.4.2
decorator==5.1.1
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
Jinja2==3.0.3
MarkupSafe==2.1.1
ratelim==0.1.6
markdown2==2.4.3
django-admin-thumbnails==0.2.6
django_jinja
channels==4.0.0
channels-redis==4.0.0
daphne
django-celery-beat==2.3.0
django-celery-results>=2.2.0
django-star-ratings==0.9.2
django-jalali-date==1.0.1
pytest==7.3.1